requirements.txt
# Airflow core y proveedores
apache-airflow==2.7.3
apache-airflow-providers-postgres==5.10.0

# DB/ORM
psycopg2-binary==2.9.9
SQLAlchemy==1.4.52  # Airflow 2.7 es más estable con 1.4.x
alembic==1.12.1

# Data
pandas==2.1.3

# APIs
requests==2.31.0
httpx==0.25.2

# Utilidades
python-dotenv==1.0.0
pytz==2023.3
pydantic==2.5.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
responses==0.24.1

# Machine Learning
scikit-learn==1.3.2
joblib==1.3.2
